add more test coverage to useCheckPublishCompleted.ts
